Use fixtures for Unit tests
