+++ tests/Unit/Fixtures/FakeEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Unit\Fixtures;
+
+class FakeEvent
+{
+    public function tags()
+    {
+        return [
+            'eventTag1',
+            'eventTag2',
+        ];
+    }
+}
