+++ tests/Unit/Fixtures/FakeEventWithModel.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Unit\Fixtures;
+
+class FakeEventWithModel
+{
+    public $model;
+
+    public function __construct($id)
+    {
+        $this->model = new FakeModel;
+        $this->model->id = $id;
+    }
+}
