+++ tests/Unit/Fixtures/FakeJobWithEloquentCollection.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Unit\Fixtures;
+
+class FakeJobWithEloquentCollection
+{
+    public $collection;
+
+    public function __construct($collection)
+    {
+        $this->collection = $collection;
+    }
+}
