+++ tests/Unit/Fixtures/FakeJobWithEloquentModel.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Unit\Fixtures;
+
+class FakeJobWithEloquentModel
+{
+    public $nonModel;
+    public $first;
+    public $second;
+
+    public function __construct($first, $second)
+    {
+        $this->nonModel = 1;
+        $this->first = $first;
+        $this->second = $second;
+    }
+}
