+++ tests/Unit/Fixtures/FakeJobWithTagsMethod.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Unit\Fixtures;
+
+class FakeJobWithTagsMethod
+{
+    public function tags()
+    {
+        return [
+            'first',
+            'second',
+        ];
+    }
+}
