+++ tests/Unit/Fixtures/FakeListener.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Unit\Fixtures;
+
+class FakeListener
+{
+    public function tags()
+    {
+        return [
+            'listenerTag1',
+            'listenerTag2',
+        ];
+    }
+}
