+++ tests/Unit/Fixtures/FakeModel.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Unit\Fixtures;
+
+use Illuminate\Database\Eloquent\Model;
+
+class FakeModel extends Model
+{
+    //
+}
